+++ samples/emoji-search/presenters/src/commonMain/kotlin/app/cash/zipline/samples/emojisearch/EmojiSearchPresenter.kt
@@ -15,39 +15,11 @@
  */
 package app.cash.zipline.samples.emojisearch
 
-import androidx.compose.runtime.Composable
 import app.cash.zipline.ZiplineService
-import app.cash.zipline.samples.emojisearch.EmojiSearchEvent.SearchTermEvent
-import example.schema.compose.Column
-import example.schema.compose.Image
-import example.schema.compose.ScrollableColumn
-import example.schema.compose.TextInput
 import kotlinx.coroutines.flow.Flow
 
 interface EmojiSearchPresenter : ZiplineService {
   fun produceModels(
     events: Flow<EmojiSearchEvent>,
   ): Flow<EmojiSearchViewModel>
-}
-
-@Composable
-fun EmojiSearch(
-  viewModel: EmojiSearchViewModel,
-  onEvent: (EmojiSearchEvent) -> Unit,
-) {
-  Column {
-    TextInput(
-      text = viewModel.searchTerm,
-      hint = "Search",
-      onTextChanged = { onEvent(SearchTermEvent(it)) },
-    )
-    ScrollableColumn {
-      for (image in viewModel.images) {
-        Image(
-          url = image.url,
-          label = image.label,
-        )
-      }
-    }
-  }
 }
